% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{addInLaterCytokineData}
\alias{addInLaterCytokineData}
\title{addInLaterCytokineData
gets cytokine sensitivity data and metadata in one call, adding to the table instead of creating new one}
\usage{
addInLaterCytokineData()
}
\description{
addInLaterCytokineData
gets cytokine sensitivity data and metadata in one call, adding to the table instead of creating new one
}
