% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{clusterDrugFamilyEfficacy}
\alias{clusterDrugFamilyEfficacy}
\title{selects AUC data and molecular data nand builds heatmap
of single drug and the results of the prediction}
\usage{
clusterDrugFamilyEfficacy(
  familyName = "MEK",
  meth = "RandomForest",
  data = "proteinLevels",
  doEnrich = FALSE,
  this.auc.dat = auc.dat,
  auc.thresh = 100,
  genes,
  data.mat = NULL,
  prefix = ""
)
}
\arguments{
\item{familyName}{Name of drug to select from prediction results}

\item{meth}{Method used to select predictors}

\item{data}{Type of data to be plotted}

\item{auc.thresh}{Threshold to determine if sample is sensitive}

\item{data.mat}{Data frame of data to plot}

\item{auc.dat}{AUC data from drug/samples}

\item{new.results}{Prediction data frame}
}
\value{
filename of pdf
}
\description{
selects AUC data and molecular data nand builds heatmap
of single drug and the results of the prediction
}
