% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{clusterSingleDrugEfficacy}
\alias{clusterSingleDrugEfficacy}
\title{selects AUC data and molecular data nand builds heatmap
of single drug and the results of the prediction}
\usage{
clusterSingleDrugEfficacy(
  drugName = "Doramapimod (BIRB 796)",
  meth = "RandomForest",
  data = "proteinLevels",
  auc.dat = auc.dat,
  auc.thresh = 100,
  new.results = new.results,
  data.mat = NULL
)
}
\arguments{
\item{drugName}{Name of drug to select from prediction results}

\item{meth}{Method used to select predictors}

\item{data}{Type of data to be plotted}

\item{auc.dat}{AUC data from drug/samples}

\item{auc.thresh}{Threshold to determine if sample is sensitive}

\item{new.results}{Prediction data frame}

\item{data.mat}{Data frame of data to plot}
}
\value{
filename of pdf
}
\description{
selects AUC data and molecular data nand builds heatmap
of single drug and the results of the prediction
}
