% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{combForest}
\alias{combForest}
\title{combForest
Runs random forest on combination of feature types}
\usage{
combForest(
  tab,
  feature.list = c("proteinLevels", "mRNAlevels", "geneMutations")
)
}
\value{
a data frame with 3 values
}
\description{
combForest
Runs random forest on combination of feature types
}
