% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{combForestEval}
\alias{combForestEval}
\title{combForest
Runs random forest on combination of feature types}
\usage{
combForestEval(
  tab,
  feature.list = c("proteinLevels", "mRNAlevels", "geneMutations")
)
}
\value{
a data frame with 3 values
}
\description{
combForest
Runs random forest on combination of feature types
}
