% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{computeDiffExByDrug}
\alias{computeDiffExByDrug}
\title{Nots ure if this is used yet
Currently deprecated}
\usage{
computeDiffExByDrug(sens.data)
}
\description{
Nots ure if this is used yet
Currently deprecated
}
