% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{computeFCDistances}
\alias{computeFCDistances}
\title{compute differences between gene lists}
\usage{
computeFCDistances(proteins.with.lfc)
}
\arguments{
\item{protein}{FC data frame}
}
\value{
genes with distance values
}
\description{
compute differences between gene lists
}
