% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{computeGSEA}
\alias{computeGSEA}
\title{compute gene set enrichment}
\usage{
computeGSEA(genes.with.values)
}
\arguments{
\item{output}{of genes and difference values}
}
\value{
gSEA output type stuff
}
\description{
compute gene set enrichment
}
