% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{computeKSEA}
\alias{computeKSEA}
\title{compute kinase substrate enrichment - osama's code wrapped in package.}
\usage{
computeKSEA(genes.with.values, ksea_FDR = 0.05, prefix = "")
}
\arguments{
\item{genes.with.values}{of genes and difference values}

\item{prot.univ}{the space of all proteins we are considering}
}
\value{
KSEA output type stuff
* KSEA using not only the known substrates in PSP but also the predicted substrates in NetworKIN
}
\description{
compute kinase substrate enrichment - osama's code wrapped in package.
}
\author{
Osama
}
