% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{computeNetworkDifferences}
\alias{computeNetworkDifferences}
\title{compute differences between networks}
\usage{
computeNetworkDifferences(network1, network2)
}
\arguments{
\item{network}{2}
}
\value{
genes and list of distance values
}
\description{
compute differences between networks
}
