% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinValues.R
\name{computePhosphoChanges}
\alias{computePhosphoChanges}
\title{get ordered phospho sites}
\usage{
computePhosphoChanges(phosData)
}
\value{
data table of log fold change
}
\description{
get ordered phospho sites
}
