% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{computeProteinFC}
\alias{computeProteinFC}
\title{compute protein FC}
\usage{
computeProteinFC(tidied.df)
}
\arguments{
\item{tidied}{data frame}

\item{condition}{of interest}
}
\value{
data frame with 3 columns - genes with LFC values and significance values
}
\description{
compute protein FC
}
