% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{createCommunityGraph}
\alias{createCommunityGraph}
\title{takes multiple lists of proteins and creates networks
then merges them to find a single community}
\usage{
createCommunityGraph(protLists, nrand, beta)
}
\arguments{
\item{protLists}{lists of proteins by condition}
}
\value{
shared graph
}
\description{
takes multiple lists of proteins and creates networks
then merges them to find a single community
}
