% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{drugMolLogReg}
\alias{drugMolLogReg}
\title{drugMolLogReg
Computes logistic regression based on AUC threshold of 100}
\usage{
drugMolLogReg(
  clin.data,
  mol.data,
  mol.feature,
  category = "Condition",
  aucThresh = 100
)
}
\description{
drugMolLogReg
Computes logistic regression based on AUC threshold of 100
}
