% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{drugMolLogRegEval}
\alias{drugMolLogRegEval}
\title{drugMolLogRegEval
Computes logistic regression and evaluates on test dataset for a particular aucThreshold}
\usage{
drugMolLogRegEval(
  clin.data,
  mol.data,
  mol.feature = "Gene",
  mol.feature.name,
  test.clin,
  test.mol,
  category = "Condition",
  aucThresh = 100
)
}
\arguments{
\item{clin.data}{AUC data for training}

\item{mol.data}{molecular data table}

\item{mol.feature}{list of molecular features to evaluate}

\item{mol.feature.name}{list of molecular feature names}

\item{test.clin}{AUC data for testing}

\item{test.mol}{molecular data for testing}

\item{category}{Column in AUC data to use}

\item{aucThresh}{Threshold to call sample sensitive vs. resistant}
}
\value{

}
\description{
drugMolLogRegEval
Computes logistic regression and evaluates on test dataset for a particular aucThreshold
}
