% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{drugMolRegressionEval}
\alias{drugMolRegressionEval}
\title{working on this function still
The goal is to use a basic elastic net regression to identify how 
well each molecular feature predicts outcome as well as how many features
are selected}
\usage{
drugMolRegressionEval(
  clin.data,
  mol.data,
  mol.feature,
  test.clin,
  test.mol,
  category = "Condition",
  doEnet = FALSE
)
}
\arguments{
\item{clin.data}{is tidied clinical data}

\item{mol.data}{is tidied table of molecular data}

\item{mol.feature}{is name of column to select from mol.data}

\item{category}{can be either Condition or family}
}
\description{
working on this function still
The goal is to use a basic elastic net regression to identify how 
well each molecular feature predicts outcome as well as how many features
are selected
}
