% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{findCandidates}
\alias{findCandidates}
\title{The goal is to pre filter features using limma, with the hope that
the smaller set of features consist of features which are more likely
to produce good models by glmnet.}
\usage{
findCandidates(mat, tmat, yvar, tyvar, p.val = 0.001)
}
\arguments{
\item{mat}{matrix samples as rows, features as columns}

\item{tmat}{see mat}

\item{yvar}{(auc) response. This is what we eventually want to predict, but here we use it to define categorical variables and train a model.}

\item{tyvar}{see yvar}

\item{p.val}{cutoff when filtering}
}
\description{
The goal is to pre filter features using limma, with the hope that
the smaller set of features consist of features which are more likely
to produce good models by glmnet.
}
