% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinValues.R
\name{getCoClusteredProteins}
\alias{getCoClusteredProteins}
\title{getCoClusteredProteins
Designed to evaluate time-course data
Gets groups of proteins that are clustered across time points in a condition of interest}
\usage{
getCoClusteredProteins(timeCourseData)
}
\value{
dataFrame
}
\description{
getCoClusteredProteins
Designed to evaluate time-course data
Gets groups of proteins that are clustered across time points in a condition of interest
}
