% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinValues.R
\name{getCorrelatedProteins}
\alias{getCorrelatedProteins}
\title{Get correlated proteins
Computes protein correlations with dose response metric, either AUC or IC50
returns list of drugs, proteins and correlation values}
\usage{
getCorrelatedProteins(data, metric = "AUC")
}
\value{
data.frame
}
\description{
Get correlated proteins
Computes protein correlations with dose response metric, either AUC or IC50
returns list of drugs, proteins and correlation values
}
