% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{getCytokineSensData}
\alias{getCytokineSensData}
\title{getCytokine data
gets cytokine sensitivity data and metadata in one call}
\usage{
getCytokineSensData()
}
\description{
getCytokine data
gets cytokine sensitivity data and metadata in one call
}
