% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{getFeaturesFromString}
\alias{getFeaturesFromString}
\title{getFeaturesFromString - separates out comma-delimited model features}
\usage{
getFeaturesFromString(string, prefix = "mRNALevels")
}
\value{
list of features
}
\description{
getFeaturesFromString - separates out comma-delimited model features
}
