% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{getFeaturesFromStringdf}
\alias{getFeaturesFromStringdf}
\title{getFeaturesFromStringdf - separates out comma-delimited model features and returns data frame}
\usage{
getFeaturesFromStringdf(string, prefix = "mRNALevels")
}
\value{
list of features
}
\description{
getFeaturesFromStringdf - separates out comma-delimited model features and returns data frame
}
