% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{getPatientDrugResponses}
\alias{getPatientDrugResponses}
\title{this grabs drug responses from three different sources, all on synapse}
\usage{
getPatientDrugResponses(patientlist)
}
\arguments{
\item{patientlist}{it will only return drugs for these patients}
}
\value{
data frame of drug response values
}
\description{
this grabs drug responses from three different sources, all on synapse
}
