% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{getSorafenibSamples}
\alias{getSorafenibSamples}
\title{getSorafenibSamples
Gets sorafenib trreated samples and stores to synaspe tables}
\usage{
getSorafenibSamples()
}
\description{
getSorafenibSamples
Gets sorafenib trreated samples and stores to synaspe tables
}
