% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{limmaTwoFactorDEAnalysis}
\alias{limmaTwoFactorDEAnalysis}
\title{limmaTwoFactorDEAnalysis
uses Osama's code to compute de from limma}
\usage{
limmaTwoFactorDEAnalysis(dat, sampleIDs.group1, sampleIDs.group2)
}
\arguments{
\item{data}{matrix}

\item{group1}{ids}

\item{group2}{ids}
}
\description{
limmaTwoFactorDEAnalysis
uses Osama's code to compute de from limma
}
\author{
Osama
}
