% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beatAMLdata.R
\name{loadBeatAMLClinicalDrugData}
\alias{loadBeatAMLClinicalDrugData}
\title{looadBeatAMLClinicalDrugData
This function looads the clinical parameters and the
drug dosage data for each patient}
\usage{
loadBeatAMLClinicalDrugData(threshold = 0.1)
}
\arguments{
\item{threshold}{is the fraction of patients with an AUC under
100, required to be included}
}
\description{
looadBeatAMLClinicalDrugData
This function looads the clinical parameters and the
drug dosage data for each patient
}
