% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beatAMLdata.R
\name{loadBeatAMLData}
\alias{loadBeatAMLData}
\title{loadBeatAMLData
General function that calls all beatAML data into memory, required for
analysis code to work}
\usage{
loadBeatAMLData()
}
\description{
loadBeatAMLData
General function that calls all beatAML data into memory, required for
analysis code to work
}
