% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinValues.R
\name{manualDEAnalysis}
\alias{manualDEAnalysis}
\title{compute manual lfc}
\usage{
manualDEAnalysis(data, group1, group2)
}
\arguments{
\item{data}{matrix}

\item{group1}{ids}

\item{group2}{ids}
}
\value{
data frame with fold change, p-value
}
\description{
compute manual lfc
}
