% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinValues.R
\name{mapPhosphoToKinase}
\alias{mapPhosphoToKinase}
\title{mapPhosphoToKInase
reads in phospho data, looads of KSDB and then computes log fold change for each kinase}
\usage{
mapPhosphoToKinase(pat.phos)
}
\value{
data table of merged data
}
\description{
mapPhosphoToKInase
reads in phospho data, looads of KSDB and then computes log fold change for each kinase
}
