% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{miniForestEval}
\alias{miniForestEval}
\title{miniForest
Runds random forest on the table and molecular feature of interest}
\usage{
miniForestEval(tab, mol.feature, quant = 0.995)
}
\value{
list
}
\description{
miniForest
Runds random forest on the table and molecular feature of interest
}
