% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{miniLogR}
\alias{miniLogR}
\title{miniDE
Carries out differential expressiona nalysis using an AUC of 100}
\usage{
miniLogR(tab, mol.feature, feature.val)
}
\arguments{
\item{mol.features}{}
}
\description{
miniDE
Carries out differential expressiona nalysis using an AUC of 100
}
