% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{miniLogREval}
\alias{miniLogREval}
\title{miniLogReval: Assesses the performance of building a logistic regression model on one test of 
data `trainTab` and assessing it on a second `testTab`}
\usage{
miniLogREval(trainTab, testTab, mol.feature)
}
\arguments{
\item{trainTab}{- a tabular form of data with three columns}

\item{testTab}{- a tabular form of data with three columns}

\item{mol.features}{}
}
\description{
miniLogReval: Assesses the performance of building a logistic regression model on one test of 
data `trainTab` and assessing it on a second `testTab`
}
