% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beatAMLdata.R
\name{plotAllPatients}
\alias{plotAllPatients}
\title{Specialized set of commands that specifically load beatAML
data - it's pretty big
summarize dataset}
\usage{
plotAllPatients(auc.data, pat.data, pphos)
}
\arguments{
\item{auc.data}{- AUC data and clinical}

\item{mol.data}{-molelcular data}
}
\description{
Specialized set of commands that specifically load beatAML
data - it's pretty big
summarize dataset
}
