% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{plotCorrelationEnrichment}
\alias{plotCorrelationEnrichment}
\title{Plot using correlation enrichment from leapR package. 
A single plot is saved to the working directory}
\usage{
plotCorrelationEnrichment(
  exprs,
  geneset,
  fdr.cutoff = 0.05,
  corr.cutoff = 0.1,
  prefix,
  width = 11,
  height = 8.5,
  order.by = "Ingroup mean",
  clean.names = FALSE,
  pathway.plot.size = 3
)
}
\arguments{
\item{exprs}{A matrix of intensities with accessions as row names, along with samples in the columns.}

\item{geneset}{Pathway/Kinase database, eg ncipid, msigdb, both of which are included in leapr.}

\item{prefix}{string, used for naming the saved plots.}

\item{order.by}{This determines how the pathways are sorted. Default is pathway correlation of "Ingroup mean", but can also use "BH_pvalue" to sort by significance of the pathways.}

\item{clean.names}{Boolean, if TRUE removes the "_pathway" ending in pathway names, making the plot easier to read.}
}
\description{
Plot using correlation enrichment from leapR package. 
A single plot is saved to the working directory
}
