% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{plotCorrelationsByDrug}
\alias{plotCorrelationsByDrug}
\title{how do we visualize the correlations of each drug/gene pair?}
\usage{
plotCorrelationsByDrug(cor.res, cor.thresh)
}
\arguments{
\item{cor.thresh}{}
}
\description{
Computes drug by element correlaction and plots them in various ways
}
