% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagOfGenesEnrichment.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Plot using clusterProfiler. 3 GSEA plots are saved to the working directory, 
two of which are custom to the amlresistancenetworks package.}
\usage{
plotGSEA(
  genes.with.values,
  prefix,
  gsea_FDR = 0.05,
  pathway.plot.size = 3,
  order.by = "NES",
  height = 8.5,
  width = 11,
  term.2.gene,
  term.2.name,
  ...
)
}
\arguments{
\item{genes.with.values}{A data frame with genes as row names, along with a column named "value". Usually this value column consists of log fold changes between two groups.}

\item{prefix}{string, used for naming the saved plots.}

\item{order.by}{This determines how the GO terms are sorted. Default is normalized enrichment score "NES", but can also use "p.adjust" to sort by significance of the terms.}
}
\description{
Plot using clusterProfiler. 3 GSEA plots are saved to the working directory, 
two of which are custom to the amlresistancenetworks package.
}
