% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinValues.R
\name{plotKinDat}
\alias{plotKinDat}
\title{plotKinDat}
\usage{
plotKinDat(
  kindat,
  phosData = phosData,
  prefix = "all",
  idcol = "Sample",
  vars = c("Sample", "cellLine", "Ligand"),
  genelist = c()
)
}
\arguments{
\item{kindat}{- output of `mapPhospoToKinase`}

\item{phosData}{- tidied phospho data}

\item{prefix}{- string for file}

\item{vars}{- sample-specific}

\item{genelist}{- select genes to plot}

\item{idcol-}{name of sample identifier}
}
\description{
plotKinDat
}
