% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivityStats.R
\name{plotProtsByMetric}
\alias{plotProtsByMetric}
\title{plotPRotsByMetric
plots basal protein expression by metric for a specific list of proteins}
\usage{
plotProtsByMetric(
  sens.data,
  genelist = c("BCL2", "TRIM21", "DUSP23"),
  metric = "AUC",
  listname = "selected"
)
}
\arguments{
\item{sens.data}{formatted data}

\item{genelist}{list of gene names to plot}

\item{metric}{either AUC or IC50}

\item{listname}{list}
}
\value{
just plots
}
\description{
plotPRotsByMetric
plots basal protein expression by metric for a specific list of proteins
}
