% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{plotSingleProtein}
\alias{plotSingleProtein}
\title{plot single protein}
\usage{
plotSingleProtein(gene, df, prefix = "")
}
\arguments{
\item{df}{tidied data fraome}
}
\description{
plot single protein
}
