% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseUtil.R
\name{querySynapseTable}
\alias{querySynapseTable}
\title{query synapse table
This is how you get data from the project}
\usage{
querySynapseTable(tableid)
}
\description{
query synapse table
This is how you get data from the project
}
