% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidyMetadata}
\alias{readAndTidyMetadata}
\title{reads in metadata from excel spreadsheet and tidies it}
\usage{
readAndTidyMetadata()
}
\description{
reads in metadata from excel spreadsheet and tidies it
}
