% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidyPhosphoProtMeasures}
\alias{readAndTidyPhosphoProtMeasures}
\title{reads and tidies phosphoprotein measurements
joins with metadata
saves tidied data and returns}
\usage{
readAndTidyPhosphoProtMeasures()
}
\description{
reads and tidies phosphoprotein measurements
joins with metadata
saves tidied data and returns
}
