% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidyProtMeasures}
\alias{readAndTidyProtMeasures}
\title{reads and tidies protein measurements
joins with metadata
saves tidied data and returns}
\usage{
readAndTidyProtMeasures()
}
\description{
reads and tidies protein measurements
joins with metadata
saves tidied data and returns
}
