% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidyQuizMetadata}
\alias{readAndTidyQuizMetadata}
\title{reads in metadata from excel spreadsheet and tidies it}
\usage{
readAndTidyQuizMetadata()
}
\description{
reads in metadata from excel spreadsheet and tidies it
}
