% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidyQuizPhosphoProtMeasures}
\alias{readAndTidyQuizPhosphoProtMeasures}
\title{reads and tidies phosphoprotein measurements
joins with metadata
saves tidied data and}
\usage{
readAndTidyQuizPhosphoProtMeasures()
}
\description{
reads and tidies phosphoprotein measurements
joins with metadata
saves tidied data and
}
