% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidyQuizProtMeasures}
\alias{readAndTidyQuizProtMeasures}
\title{reads and tidies protein measurements from quizartinib treatment
joins with metadata
saves tidied data and returns}
\usage{
readAndTidyQuizProtMeasures()
}
\description{
reads and tidies protein measurements from quizartinib treatment
joins with metadata
saves tidied data and returns
}
