% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidySensPhosMeasures}
\alias{readAndTidySensPhosMeasures}
\title{reads and tidies phosphototeomic data for experimnet 11}
\usage{
readAndTidySensPhosMeasures()
}
\description{
reads and tidies phosphototeomic data for experimnet 11
}
