% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidySensProtMeasure}
\alias{readAndTidySensProtMeasure}
\title{reads and tidies bulk proteomic data for experiment 11}
\usage{
readAndTidySensProtMeasure()
}
\description{
reads and tidies bulk proteomic data for experiment 11
}
