% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{readAndTidyUncorrectedQuizPhosphoMeasures}
\alias{readAndTidyUncorrectedQuizPhosphoMeasures}
\title{reads and tidies uncorrected phosphoprotein measurements
joins with metadata
saves tidied data and returns}
\usage{
readAndTidyUncorrectedQuizPhosphoMeasures()
}
\description{
reads and tidies uncorrected phosphoprotein measurements
joins with metadata
saves tidied data and returns
}
