% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseUtil.R
\name{synapseLogin}
\alias{synapseLogin}
\title{Logs into Synapse using local information}
\usage{
synapseLogin()
}
\value{
Synapse login python entity
}
\description{
Logs into Synapse using local information
}
