% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseUtil.R
\name{synTableStore}
\alias{synTableStore}
\title{Synapse table store
REQUIRES PANDAS ON YOUR PYTHON PATH}
\usage{
synTableStore(tab, tabname, parentId = "syn22128879")
}
\arguments{
\item{tabname}{name of table}

\item{parentId}{id of project}

\item{table}{to store on synapse}
}
\description{
Synapse table store
REQUIRES PANDAS ON YOUR PYTHON PATH
}
