% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseUtil.R
\name{synTableUpdate}
\alias{synTableUpdate}
\title{synTableUpdate
Removes all rows from table and replaces}
\usage{
synTableUpdate(tab, synid)
}
\arguments{
\item{synapse}{id of table to replace}
}
\description{
synTableUpdate
Removes all rows from table and replaces
}
