% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{buildFeatureMatrix}
\alias{buildFeatureMatrix}
\title{buildFeatureMatrix
Builds a matrix for regression with rows as patients and columns as gene}
\usage{
buildFeatureMatrix(tab, mol.feature, sampname = "AML sample")
}
\value{
matrix
}
\description{
buildFeatureMatrix
Builds a matrix for regression with rows as patients and columns as gene
}
