% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{computeFoldChangePvals}
\alias{computeFoldChangePvals}
\title{Reads in data frame and computes fold change}
\usage{
computeFoldChangePvals(
  g.data,
  control = "None",
  conditions = c("FLT3", "FGF2")
)
}
\arguments{
\item{data.frame}{}
}
\value{
tidied data frame with p-values and fold change
}
\description{
Reads in data frame and computes fold change
}
