% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{computeGSEA}
\alias{computeGSEA}
\title{compute gene set enrichment}
\usage{
computeGSEA(genes.with.values, prot.univ)
}
\arguments{
\item{genes.with.values}{of genes and difference values}

\item{prot.univ}{the space of all proteins we are considering}
}
\value{
gSEA output type stuff
}
\description{
compute gene set enrichment
}
