% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomicStats.R
\name{computeGSEA}
\alias{computeGSEA}
\title{compute gene set enrichment - osama's code wrapped in package.}
\usage{
computeGSEA(genes.with.values, prefix, gsea_FDR = 0.05)
}
\arguments{
\item{genes.with.values}{of genes and difference values}

\item{prot.univ}{the space of all proteins we are considering}
}
\value{
gSEA output type stuff
* GSEA using gene ontology biological process gene sets
}
\description{
compute gene set enrichment - osama's code wrapped in package.
}
\author{
Osama
}
