% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinNetworks.R
\name{computePhosphoNetwork}
\alias{computePhosphoNetwork}
\title{computePhosphoNetwork
maps kinase activity to proteins in a sum}
\usage{
computePhosphoNetwork(
  phos.vals = c(),
  prot.vals = c(),
  gene.vals = c(),
  nrand = 100,
  beta = 2,
  fname
)
}
\arguments{
\item{phos.vals}{named list of phosphosite values}

\item{prot.vals}{named list of protein values}

\item{nrand}{number of randomizations}

\item{beta}{beta value}
}
\description{
computePhosphoNetwork
maps kinase activity to proteins in a sum
}
