% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinNetworks.R
\name{computeProteinNetwork}
\alias{computeProteinNetwork}
\title{computeProteinNetwork}
\usage{
computeProteinNetwork(all.vals, nrand = 100)
}
\arguments{
\item{nrand}{number of randomizations}

\item{data.frame}{all.vals with required values, columsn should be: `Gene`,`value`,`signif`, and `cond`}
}
\value{
file name of graph in gml format
}
\description{
computeProteinNetwork
}
