% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{drugMolLogRegEval}
\alias{drugMolLogRegEval}
\title{drugMolLogReg
Computes logistic regression based on AUC threshold of 100}
\usage{
drugMolLogRegEval(
  clin.data,
  mol.data,
  mol.feature,
  test.clin,
  test.mol,
  category = "Condition",
  aucThresh = 100
)
}
\description{
drugMolLogReg
Computes logistic regression based on AUC threshold of 100
}
