% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{drugMolRandomForest}
\alias{drugMolRandomForest}
\title{drugMolRandomForest
builds random forest predictor of molecular data}
\usage{
drugMolRandomForest(
  clin.data,
  mol.data,
  mol.feature,
  mol.feature.name,
  category = "Condition"
)
}
\description{
drugMolRandomForest
builds random forest predictor of molecular data
}
