% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{drugMolRegression}
\alias{drugMolRegression}
\title{working on this function still
The goal is to use a basic elastic net regression to identify how 
well each molecular feature predicts outcome as well as how many features
are selected}
\usage{
drugMolRegression(clin.data, mol.data, mol.feature, category = "Condition")
}
\arguments{
\item{clin.data}{is tidied clinical data}

\item{mol.data}{is tidied table of molecular data}

\item{mol.feature}{is name of column to select from mol.data}

\item{category}{can be either 'Condition' or 'family' which is how to group AUC values}

\item{tab}{with column names `AML sample`,Gene, and whatever the value of 'mol.feature' is.}
}
\value{
a data.frame with three values/colulmns: MSE, NumGenes, and Genes
}
\description{
working on this function still
The goal is to use a basic elastic net regression to identify how 
well each molecular feature predicts outcome as well as how many features
are selected
}
