% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularSigs.R
\name{miniForest}
\alias{miniForest}
\title{miniForest
Runds random forest on the table and molecular feature of interest}
\usage{
miniForest(tab, mol.feature, feature.val, quant = 0.995)
}
\value{
list
}
\description{
miniForest
Runds random forest on the table and molecular feature of interest
}
