% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{miniForestEval}
\alias{miniForestEval}
\title{miniForestEval}
\usage{
miniForestEval(tab, mol.feature, quant = 0.995)
}
\arguments{
\item{tab}{Data table of merged info}
}
\value{
list
}
\description{

}
\details{

}
