% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{miniLogREval}
\alias{miniLogREval}
\title{miniDE
Carries out differential expressiona nalysis using an AUC of 100}
\usage{
miniLogREval(trainTab, testTab, mol.feature)
}
\arguments{
\item{mol.features}{}
}
\description{
miniDE
Carries out differential expressiona nalysis using an AUC of 100
}
