% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAndTestMolSigs.R
\name{miniRegEval}
\alias{miniRegEval}
\title{miniRegEval
Runs lasso regression on a single feature from tabular data and evaluates on a second set of data.
This is a little more complicated than LOO
analysis because we have to check for shared feature names}
\usage{
miniRegEval(trainTab, testTab, mol.feature, enet.alpha = c(1))
}
\arguments{
\item{trainTab}{with column names `AML sample`,meanVal,Gene, and whatever the value of `mol.feature` is.}

\item{testTab}{with column names `Sample`, meanVal, Gene, and whatever the value of `mol.feature` is}

\item{mol.feature}{The molecular feature to be evaluated}
}
\value{
a data.frame with three values/columns: MSE, numFeatures, and Genes
}
\description{
miniRegEval
Runs lasso regression on a single feature from tabular data and evaluates on a second set of data.
This is a little more complicated than LOO
analysis because we have to check for shared feature names
}
