% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugSensitivityStats.R
\name{plotAllAUCs}
\alias{plotAllAUCs}
\title{plotall AUCS with clinical data
what are we measuring here}
\usage{
plotAllAUCs(auc.data, to.plot = "AUC")
}
\description{
plotall AUCS with clinical data
what are we measuring here
}
